Generalize leveraged optimiser and add TESLA2x
